pyproject.toml
[tool.poetry]
name = "product-classifier"
version = "0.1.0"
description = "Classification de catégorie produit"
authors = ["Votre Nom <votre@email.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pandas = "^2.0.0"
transformers = {extras = ["torch"], version = "^4.30.0"}
datasets = "^2.0.0"
accelerate = "^0.20.0"
evaluate = "^0.4.0"
pyyaml = "^6.0"
tqdm = "^4.0.0"
scikit-learn = "^1.0.0"
jupyter = "^1.1.1"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"
wordcloud = "^1.9.4"
nltk = "^3.9.1"
plotly = "^6.1.2"
notebook = "^7.4.3"
openpyxl = "^3.1.5"
sentencepiece = "^0.2.0"
torch = {version = "^2.5.1", source = "pytorch_cuda"}
torchvision = {version = "^0.20.1", source = "pytorch_cuda"}
torchaudio = {version = "^2.5.1", source = "pytorch_cuda"}
onnx = "^1.18.0"
onnxruntime-gpu = "^1.22.0"
gradio = "^5.33.2"
spacy = "^3.8.7"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
ipython = "^8.0.0"

[tool.poetry.group.dev.dependencies]
notebook = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "pytorch_cuda"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"